converted to maven project

+++ src/main/java/org/example/backend/service/DragonRepository.java
@@ -5,5 +5,5 @@ import org.apache.deltaspike.data.api.Repository;
 import org.example.backend.Dragon;
 
 @Repository(forEntity = Dragon.class)
-public interface DragonRepository extends EntityRepository<Dragon, Integer>{
+public interface DragonRepository extends EntityRepository<Dragon, Long>{
 }
